+++ pokemongo_bot/event_handlers/__init__.py
@@ -4,3 +4,4 @@ from .socketio_handler import SocketIoHandler
 from .social_handler import SocialHandler
 from .telegram_handler import TelegramHandler
 from .discord_handler import DiscordHandler
+from .chat_handler import ChatHandler
