+++ pokemongo_bot/event_handlers/telegram_handler.py
@@ -9,6 +9,7 @@ import pprint
 from pokemongo_bot.datastore import Datastore
 from pokemongo_bot import inventory
 from telegram.utils import request
+from chat_handler import ChatHandler
 
 DEBUG_ON = False
 
@@ -19,7 +20,7 @@ class TelegramClass:
     def __init__(self, bot, master, pokemons, config):
         self.bot = bot
         request.CON_POOL_SIZE = 16
-
+        self.chat_handler = ChatHandler(self.bot, pokemons)
         with self.bot.database as conn:
             # initialize the DB table if it does not exist yet
             initiator = TelegramDBInit(bot.database)
@@ -49,212 +50,12 @@ class TelegramClass:
         self._tbot = None
         self.config = config
 
-    def sendMessage(self, chat_id=None, parse_mode='Markdown', text=None):
-        try:
-            self._tbot.sendMessage(chat_id=chat_id, parse_mode=parse_mode, text=text)
-        except telegram.error.NetworkError:
-            time.sleep(1)
-        except telegram.error.TelegramError:
-            time.sleep(10)
-        except telegram.error.Unauthorized:
-            self.update_id += 1
-
-    def sendLocation(self, chat_id, latitude, longitude):
-        try:
-            self._tbot.send_location(chat_id=chat_id, latitude=latitude, longitude=longitude)
-        except telegram.error.NetworkError:
-            time.sleep(1)
-        except telegram.error.TelegramError:
-            time.sleep(10)
-        except telegram.error.Unauthorized:
-            self.update_id += 1
-
     def connect(self):
         self._tbot = telegram.Bot(self.bot.config.telegram_token)
         try:
             self.update_id = self._tbot.getUpdates()[0].update_id
         except IndexError:
             self.update_id = None
-
-    def _get_player_stats(self):
-        return inventory.player().player_stats
-
-    def get_evolved(self, chat_id, num, order):
-        if not num.isnumeric():
-            num = 10
-        else:
-            num = int(num)
-
-        if order not in ["cp", "iv"]:
-            order = "iv"
-
-        with self.bot.database as conn:
-            cur = conn.cursor()
-            cur.execute("SELECT * FROM evolve_log ORDER BY " + order + " ASC LIMIT " + str(num))
-            evolved = cur.fetchall()
-            if evolved:
-                for x in evolved:
-                    res = (
-                        "*"+str(x[0])+"*",
-                        "_CP:_ " + str(x[2]),
-                        "_IV:_ " + str(x[1]),
-                        str(x[3])
-                        )
-
-                    self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            else:
-                self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="No Evolutions Found.\n")
-
-    def get_softban(self, chat_id):
-        with self.bot.database as conn:
-            cur = conn.cursor()
-            cur.execute("SELECT * FROM softban_log")
-            softban = cur.fetchall()
-            if softban:
-                for x in softban:
-                    res = (
-                        "*" + str(x[0]) + "*",
-                        str(x[2]))
-                    self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            else:
-                self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="No Softbans found! Good job!\n")
-
-    def get_hatched(self, chat_id, num, order):
-        if not num.isnumeric():
-            num = 10
-        else:
-            num = int(num)
-
-        if order not in ["cp", "iv"]:
-            order = "iv"
-        with self.bot.database as conn:
-            cur = conn.cursor()
-            cur.execute("SELECT * FROM eggs_hatched_log ORDER BY " + order + " ASC LIMIT " + str(num))
-            hatched = cur.fetchall()
-            if hatched:
-                for x in hatched:
-                    res = (
-                        "*" + str(x[0]) + "*",
-                        "_CP:_ " + str(x[1]),
-                        "_IV:_ " + str(x[2]),
-                        str(x[4])
-                        )
-                    self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            else:
-                self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="No Eggs Hatched Yet.\n")
-
-    def get_caught(self, chat_id, num, order):
-        if not num.isnumeric():
-            num = 10
-        else:
-            num = int(num)
-
-        if order not in ["cp", "iv"]:
-            order = "iv"
-
-        with self.bot.database as conn:
-            cur = conn.cursor()
-            cur.execute("SELECT * FROM catch_log ORDER BY " + order + " ASC LIMIT " + str(num))
-            caught = cur.fetchall()
-            if caught:
-                for x in caught:
-                    res = (
-                        "*" + str(x[0]) + "*",
-                        "_CP:_ " + str(x[1]),
-                        "_IV:_ " + str(x[2]),
-                        str(x[5])
-                        )
-                    self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            else:
-                self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="No Pokemon Caught Yet.\n")
-
-    def get_pokestops(self, chat_id, num):
-        with self.bot.database as conn:
-            cur = conn.cursor()
-            cur.execute("SELECT * FROM pokestop_log ORDER BY dated ASC LIMIT " + str(num))
-            pokestop = cur.fetchall()
-            if pokestop:
-                for x in pokestop:
-                    res = (
-                        "*" + str(x[0] + "*"),
-                        "_XP:_ " + str(x[1]),
-                        "_Items:_ " + str(x[2]),
-                        str(x[3])
-                    )
-                    self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            else:
-                self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="No Pokestops Encountered Yet.\n")
-
-    def get_transfers(self, chat_id, num, order):
-        if not num.isnumeric():
-            num = 10
-        else:
-            num = int(num)
-
-        if order not in ["cp", "iv"]:
-            order = "iv"
-        with self.bot.database as conn:
-            cur = conn.cursor()
-            cur.execute("SELECT * FROM transfer_log ORDER BY " + order + " ASC LIMIT " + str(num))
-            transfer = cur.fetchall()
-            if transfer:
-                for x in transfer:
-                    res = (
-                        "*" + str(x[0]) + "*",
-                        "_CP:_ " + str(x[2]),
-                        "_IV:_ " + str(x[1]),
-                        str(x[3])
-                    )
-                    self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            else:
-                self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="No Pokemon Released Yet.\n")
-
-    def get_vanished(self, chat_id, num, order):
-        if not num.isnumeric():
-            num = 10
-        else:
-            num = int(num)
-
-        if order not in ["cp", "iv"]:
-            order = "iv"
-        with self.bot.database as conn:
-            cur = conn.cursor()
-            cur.execute("SELECT * FROM vanish_log ORDER BY " + order + " ASC LIMIT " + str(num))
-            vanished = cur.fetchall()
-            if vanished:
-                for x in vanished:
-                    res = (
-                        "*" + str(x[0]) + "*",
-                        "_CP:_ " + str(x[1]),
-                        "_IV:_ " + str(x[2]),
-                        "_NCP:_ " + str(x[4]),
-                        str(x[5])
-                    )
-                    self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            else:
-                self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="No Pokemon Vanished Yet.\n")
-
-    def send_player_stats_to_chat(self, chat_id):
-        stats = self._get_player_stats()
-        if stats:
-            with self.bot.database as conn:
-                cur = conn.cursor()
-                cur.execute("SELECT DISTINCT COUNT(encounter_id) FROM catch_log WHERE dated >= datetime('now','-1 day')")
-                catch_day = cur.fetchone()[0]
-                cur.execute("SELECT DISTINCT COUNT(pokestop) FROM pokestop_log WHERE dated >= datetime('now','-1 day')")
-                ps_day = cur.fetchone()[0]
-                res = (
-                    "*"+self.bot.config.username+"*",
-                    "_Level:_ "+str(stats["level"]),
-                    "_XP:_ "+str(stats["experience"])+"/"+str(stats["next_level_xp"]),
-                    "_Pokemons Captured:_ "+str(stats["pokemons_captured"])+" ("+str(catch_day)+" _last 24h_)",
-                    "_Poke Stop Visits:_ "+str(stats["poke_stop_visits"])+" ("+str(ps_day)+" _last 24h_)",
-                    "_KM Walked:_ "+str("%.2f" % stats["km_walked"])
-                )
-            self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="\n".join(res))
-            self.sendLocation(chat_id=chat_id, latitude=self.bot.api._position_lat, longitude=self.bot.api._position_lng)
-        else:
-            self.sendMessage(chat_id=chat_id, parse_mode='Markdown', text="Stats not loaded yet\n")
 
     def grab_uid(self, update):
         with self.bot.database as conn:
@@ -291,70 +92,20 @@ class TelegramClass:
             cur = conn.cursor()
             cur.execute("delete from telegram_logins where uid = ?", [update.message.chat_id])
             conn.commit()
-        self.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Logout completed")
+        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Logout completed")
         return
 
     def authenticate(self, update):
         (command, password) = update.message.text.split(' ')
         if password != self.config.get('password', None):
-            self.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Invalid password")
+            self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Invalid password")
         else:
             with self.bot.database as conn:
                 cur = conn.cursor()
                 cur.execute("delete from telegram_logins where uid = ?", [update.message.chat_id])
                 cur.execute("insert into telegram_logins(uid) values(?)", [update.message.chat_id])
                 conn.commit()
-            self.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Authentication successful, you can now use all commands")
-        return
-
-    def display_events(self, update):
-        cmd = update.message.text.split(" ", 1)
-        if len(cmd) > 1:
-            # we have a filter
-            event_filter = ".*{}-*".format(cmd[1])
-        else:
-            # no filter
-            event_filter = ".*"
-        events = filter(lambda k: re.match(event_filter, k), self.bot.event_manager._registered_events.keys())
-        events.remove('vanish_log')
-        events.remove('eggs_hatched_log')
-        events.remove('catch_log')
-        events.remove('pokestop_log')
-        events.remove('load_cached_location')
-        events.remove('location_cache_ignored')
-        events.remove('softban_log')
-        events.remove('loaded_cached_forts')
-        events.remove('login_log')
-        events.remove('evolve_log')
-        events.remove('transfer_log')
-        events.remove('catchable_pokemon')
-        self.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML', text=("\n".join(events)))
-
-    def showtop(self, chatid, num, order):
-        if not num.isnumeric():
-            num = 10
-        else:
-            num = int(num)
-
-        if order not in ["cp", "iv"]:
-            order = "iv"
-
-        pkmns = sorted(inventory.pokemons().all(), key=lambda p: getattr(p, order), reverse=True)[:num]
-
-        outMsg = "\n".join(["<b>{}</b> \nCP: {} \nIV: {} \nCandy: {}\n".format(p.name, p.cp, p.iv, inventory.candies().get(p.pokemon_id).quantity) for p in pkmns])
-        self.sendMessage(chat_id=chatid, parse_mode='HTML', text=outMsg)
-
-        return
-
-
-    def evolve(self, chatid, uid):
-        # TODO: here comes evolve logic (later)
-        self.sendMessage(chat_id=chatid, parse_mode='HTML', text="Evolve logic not implemented yet")
-        return
-
-    def upgrade(self, chatid, uid):
-        # TODO: here comes upgrade logic (later)
-        self.sendMessage(chat_id=chatid, parse_mode='HTML', text="Upgrade logic not implemented yet")
+            self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Authentication successful, you can now use all commands")
         return
 
     def run(self):
@@ -380,15 +131,15 @@ class TelegramClass:
                             "/hatched <num> <cp-or-iv> - show top x pokemon hatched, sorted by CP or IV",
                             "/caught <num> <cp-or-iv>- show top x pokemon caught, sorted by CP or IV",
                             "/pokestops - show last x pokestops visited",
-                            "/transfers <num> <cp-or-iv> - show top x transfers, sorted by CP or IV",
+                            "/released <num> <cp-or-iv> - show top x released, sorted by CP or IV",
                             "/vanished <num> <cp-or-iv> - show top x vanished, sorted by CP or IV",
                             "/softbans - info about possible softbans"
                         )
-                        self.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="\n".join(res))
+                        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="\n".join(res))
                         continue
 
                     if self.config.get('password', None) == None and (not hasattr(self, "master") or not self.config.get('master', None)): # no auth provided in config
-                        self.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="No password nor master configured in TelegramTask section, bot will not accept any commands")
+                        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="No password nor master configured in TelegramTask section, bot will not accept any commands")
                         continue
                     if re.match(r'^/login [^ ]+', update.message.text):
                         self.authenticate(update)
@@ -410,69 +161,72 @@ class TelegramClass:
                         # Reject message if sender does not match defined master in config
                         outMessage = "Telegram message received from unknown sender. Please either make sure your username or ID is in TelegramTask/master, or a password is set in TelegramTask section and /login is issued"
                         self.bot.logger.error(outMessage)
-                        self.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Please /login first")
+                        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Please /login first")
                         continue
                     # one way or another, the user is now authenticated
                     if update.message.text == "/info":
-                        self.send_player_stats_to_chat(update.message.chat_id)
+                        self.chat_handler.send_player_stats_to_chat(update.message.chat_id)
                         continue
                     if update.message.text == "/softbans":
-                        self.get_softban(update.message.chat_id)
+                        self.chat_handler.get_softban(update.message.chat_id)
                         continue
                     if re.match("^/events", update.message.text):
-                        self.display_events(update)
+                        events = self.chat_handler.get_events(update)
+                        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML',
+                                                      text="\n".join(events))
                         continue
                     if update.message.text == "/logout":
-                        self.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML', text=("Logged out."))
+                        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML', text=("Logged out."))
                         self.deauthenticate(update)
                         continue
                     if re.match(r'^/sub ', update.message.text):
                         self.chsub(update.message.text, update.message.chat_id)
-                        self.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML', text=("Subscriptions updated."))
+                        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML',
+                                                      text=("Subscriptions updated."))
                         continue
                     if re.match(r'^/unsub ', update.message.text):
                         self.chsub(update.message.text, update.message.chat_id)
-                        self.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML', text=("Subscriptions updated."))
+                        self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='HTML', text=("Subscriptions updated."))
                         continue
                     if re.match(r'^/showsubs', update.message.text):
                         self.showsubs(update.message.chat_id)
                         continue
                     if re.match(r'^/top ', update.message.text):
                         (cmd, num, order) = self.tokenize(update.message.text, 3)
-                        self.showtop(update.message.chat_id, num, order)
+                        self.chat_handler.showtop(update.message.chat_id, num, order)
                         continue
                     if re.match(r'^/caught ', update.message.text):
                         (cmd, num, order) = self.tokenize(update.message.text, 3)
-                        self.get_caught(update.message.chat_id, num, order)
+                        self.chat_handler.get_caught(update.message.chat_id, num, order)
                         continue
                     if re.match(r'^/evolved ', update.message.text):
                         (cmd, num, order) = self.tokenize(update.message.text, 3)
-                        self.get_evolved(update.message.chat_id, num, order)
+                        self.chat_handler.get_evolved(update.message.chat_id, num, order)
                         continue
                     if re.match(r'^/pokestops ', update.message.text):
                         (cmd, num) = self.tokenize(update.message.text, 2)
-                        self.get_pokestops(update.message.chat_id, num)
+                        self.chat_handler.get_pokestops(update.message.chat_id, num)
                         continue
                     if re.match(r'^/hatched ', update.message.text):
                         (cmd, num, order) = self.tokenize(update.message.text, 3)
-                        self.get_hatched(update.message.chat_id, num, order)
+                        self.chat_handler.get_hatched(update.message.chat_id, num, order)
                         continue
-                    if re.match(r'^/transfers ', update.message.text):
+                    if re.match(r'^/released ', update.message.text):
                         (cmd, num, order) = self.tokenize(update.message.text, 3)
-                        self.get_transfers(update.message.chat_id, num, order)
+                        self.chat_handler.get_released(update.message.chat_id, num, order)
                         continue
                     if re.match(r'^/vanished ', update.message.text):
                         (cmd, num, order) = self.tokenize(update.message.text, 3)
-                        self.get_vanished(update.message.chat_id, num, order)
+                        self.chat_handler.get_vanished(update.message.chat_id, num, order)
                         continue
-                    self.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Unrecognized command: {}".format(update.message.text))
+                    self.chat_handler.sendMessage(chat_id=update.message.chat_id, parse_mode='Markdown', text="Unrecognized command: {}".format(update.message.text))
 
     def showsubs(self, chatid):
         subs = []
         with self.bot.database as conn:
             for sub in conn.execute("select uid, event_type, parameters from telegram_subscriptions where uid = ?", [chatid]).fetchall():
                 subs.append("{} -&gt; {}".format(sub[1], sub[2]))
-        self.sendMessage(chat_id=chatid, parse_mode='HTML', text="\n".join(subs))
+        self.chat_handler.sendMessage(chat_id=chatid, parse_mode='HTML', text="\n".join(subs))
 
     def chsub(self, msg, chatid):
         (cmd, evt, params) = self.tokenize(msg, 3)
@@ -533,11 +287,13 @@ class TelegramHandler(EventHandler):
         self.pokemons = config.get('alert_catch', {})
         self.whoami = "TelegramHandler"
         self.config = config
+        self.chat_handler = ChatHandler(self.bot, self.pokemons)
         if master == None:
             self.master = None
             return
         else:
             self.master = master
+
         with self.bot.database as conn:
             # if master is not numeric, try to fetch it from the database
             if not unicode(master).isnumeric():
@@ -581,24 +337,10 @@ class TelegramHandler(EventHandler):
                 self.tbot = None
                 self.bot.logger.error("Unable to start Telegram bot; master: {}, exception: {}".format(selfmaster, pprint.pformat(inst)))
                 return
-        try:
-            # prepare message to send
-            if event == 'level_up':
-                msg = "level up ({})".format(data["current_level"])
-            elif event == 'pokemon_caught':
-                msg = "Caught {} CP: {}, IV: {}".format(data["pokemon"], data["cp"], data["iv"])
-            elif event == 'egg_hatched':
-                msg = "Egg hatched with a {} CP: {}, IV: {} {}".format(data["name"], data["cp"], data["iv_ads"], data["iv_pct"])
-            elif event == 'bot_sleep':
-                msg = "I am too tired, I will take a sleep till {}.".format(data["wake"])
-            elif event == 'catch_limit':
-                msg = "*You have reached your daily catch limit, quitting.*"
-            elif event == 'spin_limit':
-                msg = "*You have reached your daily spin limit, quitting.*"
-            else:
-                msg = formatted_msg
-        except KeyError:
-            msg = "Error on event {}".format(event)
+            msg = self.chat_handler.get_event(event, formatted_msg, data)
+            if msg is None:
+                return
+
         # first handle subscriptions; they are independent of master setting.
         with self.bot.database as conn:
             subs = conn.execute("select uid, parameters, event_type from telegram_subscriptions where event_type in (?,'all','debug')", [event]).fetchall()
@@ -622,7 +364,9 @@ class TelegramHandler(EventHandler):
                         self.bot.logger.info("[{}] {}".format(event, msg))
 
                     else:
-                        self.tbot.sendMessage(chat_id=uid, parse_mode='Markdown', text=msg)
+                        msg = self.chat_handler.get_event(event, formatted_msg, data)
+                        if msg is None:
+                            return
 
         if hasattr(self, "master") and self.master:
             if not unicode(self.master).isnumeric():
@@ -630,19 +374,5 @@ class TelegramHandler(EventHandler):
                 # cannot send event notifications to non-numeric master (yet), so quitting
                 return
             master = self.master
-
-            if event == 'level_up':
-                msg = "level up ({})".format(data["current_level"])
-            elif event == 'egg_hatched':
-                msg = "Egg hatched with a {} CP: {}, IV: {} {}".format(data["name"], data["cp"], data["iv_ads"], data["iv_pct"])
-            elif event == 'bot_sleep':
-                msg = "I am too tired, I will take a sleep till {}.".format(data["wake"])
-            elif event == 'catch_limit':
-                self.tbot.send_player_stats_to_chat(master)
-                msg = "*You have reached your daily catch limit, quitting.*"
-            elif event == 'spin_limit':
-                self.tbot.send_player_stats_to_chat(master)
-                msg = "*You have reached your daily spin limit, quitting.*"
-            else:
-                return
-            self.tbot.sendMessage(chat_id=master, parse_mode='Markdown', text=msg)
+            msg = self.chat_handler.get_event(event, formatted_msg, data)
+            self.chat_handler.sendMessage(chat_id=master, parse_mode='Markdown', text=msg)
